Extract abstract base
